pyproject.toml
[tool.poetry]
name = "image-gallery"
version = "0.1.0"
description = "A functional demo app for the Invisible Screen"
authors = ["Invisible Computers Gmbh <info@invisible-computers.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "image_gallery"}]

[tool.poetry.dependencies]
python = "3.13.*"
django = "4.2.*"
dj-database-url = "^1.3.0"
psycopg2-binary = "^2.9.6"
pyjwt = "^2.6.0"
pytz = "^2023.3"
requests = "^2.28.2"
djangorestframework = "^3.14.0"
whitenoise = "^6.4.0"
uwsgi = "^2.0.21"
cryptography = "^43.0.3"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]

[tool.ruff]
ignore = ["E501"]
select = ["E", "F", "I"]
